refactor: move clock icon to icons folder
